Backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0

sqlalchemy==2.0.44
PyMySQL==1.1.2

pydantic==2.11.7
pydantic-settings==2.11.0
email-validator==2.2.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

redis==5.2.1
requests==2.32.3